backend/requirements.txt
fastapi
uvicorn
python-dotenv
nltk
google-generativeai